refactor workspaces
